+++ code/src/sixsq/nuvla/server/resources/module_applications_sets.clj
@@ -1,0 +1,105 @@
+(ns sixsq.nuvla.server.resources.module-applications-sets
+  "
+This resource represents applications sets module.
+"
+  (:require
+    [sixsq.nuvla.server.resources.common.crud :as crud]
+    [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
+    [sixsq.nuvla.server.resources.common.utils :as u]
+    [sixsq.nuvla.server.resources.resource-metadata :as md]
+    [sixsq.nuvla.server.resources.spec.module-applications-sets :as spec]
+    [sixsq.nuvla.server.util.metadata :as gen-md]))
+
+
+(def ^:const resource-type (u/ns->type *ns*))
+
+
+(def ^:const collection-type (u/ns->collection-type *ns*))
+
+
+(def collection-acl {:query ["group/nuvla-admin"]
+                     :add   ["group/nuvla-admin"]})
+
+
+(def resource-acl {:owners ["group/nuvla-admin"]})
+
+
+;;
+;; multimethods for validation and operations
+;;
+
+(def validate-fn (u/create-spec-validation-fn ::spec/schema))
+(defmethod crud/validate resource-type
+  [resource]
+  (validate-fn resource))
+
+
+;;
+;; use default ACL method
+;;
+
+(defmethod crud/add-acl resource-type
+  [resource _request]
+  (assoc resource :acl resource-acl))
+
+
+;;
+;; CRUD operations
+;;
+
+(def add-impl (std-crud/add-fn resource-type collection-acl resource-type))
+
+
+(defmethod crud/add resource-type
+  [request]
+  ;; should validate apps sets names are distinct
+  ;; should check that module exists
+  ;; should check that apps sets are same type k8s/docker
+  (add-impl request))
+
+
+(def retrieve-impl (std-crud/retrieve-fn resource-type))
+
+
+(defmethod crud/retrieve resource-type
+  [request]
+  (retrieve-impl request))
+
+
+(def edit-impl (std-crud/edit-fn resource-type))
+
+
+(defmethod crud/edit resource-type
+  [request]
+  ;; should validate apps sets names are distinct
+  ;; should check that module exists
+  ;; should check that apps sets are same type k8s/docker
+  (edit-impl request))
+
+
+(def delete-impl (std-crud/delete-fn resource-type))
+
+
+(defmethod crud/delete resource-type
+  [request]
+  (delete-impl request))
+
+
+(def query-impl (std-crud/query-fn resource-type collection-acl collection-type))
+
+
+(defmethod crud/query resource-type
+  [request]
+  (query-impl request))
+
+
+;;
+;; initialization
+;;
+
+(def resource-metadata (gen-md/generate-metadata ::ns ::spec/schema))
+
+(defn initialize
+  []
+  (std-crud/initialize resource-type ::spec/schema)
+  (md/register resource-metadata))